Added tests for structural mechanics (static and dynamic), fluid dynamics, and thermal cases.
