+++ applications/RomApplication/tests/structural_dynamic_test_files/ROM/ExpectedOutput.npy
