+++ applications/RomApplication/tests/thermal_dynamic_test_files/ROM/ExpectedOutput.npy
