+++ applications/RomApplication/tests/thermal_static_test_files/LSPGROM/ProjectParameters.json
@@ -15,10 +15,10 @@
         "domain_size": 2,
         "model_import_settings": {
             "input_type": "mdpa",
-            "input_filename": "Square_Radiation_Stationary"
+            "input_filename": "../Square_Radiation_Stationary"
         },
         "material_import_settings": {
-            "materials_filename": "ConvectionDiffusionMaterials.json"
+            "materials_filename": "../ConvectionDiffusionMaterials.json"
         },
         "line_search": false,
         "echo_level": 0,
