+++ applications/RomApplication/tests/thermal_static_test_files/ROM/ExpectedOutput.npy
