+++ applications/RomApplication/tests/thermal_static_test_files/ROM/RomParameters.json
